Vue component, ES modules
