+++ team15/components/elements/buttons.js
@@ -1,0 +1,14 @@
+export const StartGameButton = {
+  name: "start-game-button",
+  template: `
+    <button class = "big-buttons" id="start-game-button">START GAME</button>
+  `
+};
+
+export const HowToPlayButton = {
+  name: "How-To-Play-Button",
+  template:`
+   <button class = "big-buttons" id="how-to-play-button">HOW TO PLAY</button>
+  `
+}
+
