+++ team15/components/index.js
@@ -1,0 +1,9 @@
+import {StartView} from  "./views/startView.js";
+import {StartGameButton, HowToPlayButton} from "./elements/buttons.js";
+
+
+export default{
+    StartGameButton,
+    HowToPlayButton,
+    StartView,
+};
