+++ team15/components/views/startView.js
