+++ team15/index.html
@@ -21,7 +21,7 @@ Owned by Team 15
   -->
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/7.0.0/css/all.min.css"> 
   <script src="https://cdn.jsdelivr.net/npm/jquery@3.7.1/dist/jquery.min.js"></script>
-  <script src="https://unpkg.com/vue@3/dist/vue.global.js"></script>
+  <!--<script src="https://unpkg.com/vue@3/dist/vue.global.js"></script>-->
 
   <!-- Site specific scripts and styles -->
   <link rel="stylesheet" href="index.css" />
@@ -33,19 +33,6 @@ Owned by Team 15
     <h1>Team 15</h1>
   </header>
 
-  <!-- TODO: Remove this div and it's content -->
-  <div>
-      <hr>
-      <h2>A basic section to make sure that everything works for everyone</h2>
-
-      <p>Font awesome is working: <i class="fa-solid fa-check"></i> (A checkmark should be visible)</p>
-
-      <button id="check-jquery">Check if Javascript and JQuery is working</button> (Clicking the button should show an alert)
-
-      <button id="check-saving">Click this button to see if saving works</button>
-      <hr>
-  </div>
-
   <main class="game-frame">
   <!-- <div class = "main-text">WELCOME TO THE DRESSING PELLE GAME</div> 
     <div class="button-wrapper">
@@ -56,7 +43,6 @@ Owned by Team 15
     <div id="app"> </div>
 
 
-    Learning: <div id="display-vocab"></div>
   </main>
 
   <footer>
